+++ app/src/fileformat.js
@@ -18,11 +18,12 @@ exports.stringify = (json) => { return JSON.stringify(json,null,2) }
 
 // Document with a single, empty cell
 exports.single    = () => { return {
-    version        : '0.2.0.0',
+    version        : '0.2.1.0',
     cells          : [{
       'cell_type' : 'code',
       source      : '',
     }],
+	extensions     : '',
     importModules  : '',
     loadFiles      : '',
     settings       : {
@@ -36,16 +37,22 @@ exports.single    = () => { return {
 // Update the file format to a new version
 exports.update    = (versionNew, input) => {
   // FIXME: Make sure that the JSON object is actually copied.
-  let output = input
+  let output     = input
+  let versionOld = input.version
 
-  if (input.version === '0.1.0.0' && versionNew === '0.2.0.0') {
-    output.version = versionNew
+  // Remark: We use lexicographic ordering to test whether we should upgrade
+  if (versionOld === '0.1.0.0' && versionNew >= '0.2.0.0') {
     output.cells   = input.cells.map( (source) => {
       return {
         'cell_type' : 'code',
         'source'    : source
       }
     })
+	versionOld     = '0.2.0.0'
+    output.version = versionOld
   }
+
+  // We can perform several upgrades in sequence.
+
   return output
 }